+++ test/models/rememberable_test.rb
@@ -55,12 +55,27 @@ class RememberableTest < ActiveSupport::TestCase
     assert resource_class.new.respond_to?(:remember_me=)
   end
 
-  test 'forget_me should clear remember_created_at' do
-    resource = create_resource
-    resource.remember_me!
-    assert_not resource.remember_created_at.nil?
-    resource.forget_me!
-    assert resource.remember_created_at.nil?
+  test 'forget_me should clear remember_created_at if expire_all_remember_me_on_sign_out is true' do
+    swap Devise, expire_all_remember_me_on_sign_out: true do
+      resource = create_resource
+      resource.remember_me!
+      assert_not_nil resource.remember_created_at
+
+      resource.forget_me!
+      assert_nil resource.remember_created_at
+    end
+  end
+
+  test 'forget_me should not clear remember_created_at if expire_all_remember_me_on_sign_out is false' do
+    swap Devise, expire_all_remember_me_on_sign_out: false do
+      resource = create_resource
+      resource.remember_me!
+
+      assert_not_nil resource.remember_created_at
+
+      resource.forget_me!
+      assert_not_nil resource.remember_created_at
+    end
   end
 
   test 'forget_me should not try to update resource if it has been destroyed' do
@@ -86,7 +101,7 @@ class RememberableTest < ActiveSupport::TestCase
   end
 
   test 'remember for should fallback to devise remember for default configuration' do
-    swap Devise, :remember_for => 1.day do
+    swap Devise, remember_for: 1.day do
       resource = create_resource
       resource.remember_me!
       assert_not resource.remember_expired?
@@ -94,7 +109,7 @@ class RememberableTest < ActiveSupport::TestCase
   end
 
   test 'remember expires at should sum date of creation with remember for configuration' do
-    swap Devise, :remember_for => 3.days do
+    swap Devise, remember_for: 3.days do
       resource = create_resource
       resource.remember_me!
       assert_equal 3.days.from_now.to_date, resource.remember_expires_at.to_date
@@ -105,7 +120,7 @@ class RememberableTest < ActiveSupport::TestCase
   end
 
   test 'remember should be expired if remember_for is zero' do
-    swap Devise, :remember_for => 0.days do
+    swap Devise, remember_for: 0.days do
       Devise.remember_for = 0.days
       resource = create_resource
       resource.remember_me!
@@ -114,7 +129,7 @@ class RememberableTest < ActiveSupport::TestCase
   end
 
   test 'remember should be expired if it was created before limit time' do
-    swap Devise, :remember_for => 1.day do
+    swap Devise, remember_for: 1.day do
       resource = create_resource
       resource.remember_me!
       resource.remember_created_at = 2.days.ago
@@ -124,7 +139,7 @@ class RememberableTest < ActiveSupport::TestCase
   end
 
   test 'remember should not be expired if it was created within the limit time' do
-    swap Devise, :remember_for => 30.days do
+    swap Devise, remember_for: 30.days do
       resource = create_resource
       resource.remember_me!
       resource.remember_created_at = (30.days.ago + 2.minutes)
@@ -134,7 +149,7 @@ class RememberableTest < ActiveSupport::TestCase
   end
 
   test 'if extend_remember_period is false, remember_me! should generate a new timestamp if expired' do
-    swap Devise, :remember_for => 5.minutes do
+    swap Devise, remember_for: 5.minutes do
       resource = create_resource
       resource.remember_me!(false)
       assert resource.remember_created_at
@@ -148,7 +163,7 @@ class RememberableTest < ActiveSupport::TestCase
   end
 
   test 'if extend_remember_period is false, remember_me! should not generate a new timestamp' do
-    swap Devise, :remember_for => 1.year do
+    swap Devise, remember_for: 1.year do
       resource = create_resource
       resource.remember_me!(false)
       assert resource.remember_created_at
@@ -162,7 +177,7 @@ class RememberableTest < ActiveSupport::TestCase
   end
 
   test 'if extend_remember_period is true, remember_me! should always generate a new timestamp' do
-    swap Devise, :remember_for => 1.year do
+    swap Devise, remember_for: 1.year do
       resource = create_resource
       resource.remember_me!(true)
       assert resource.remember_created_at